automatically backed up by learn

+++ index.js
@@ -37,5 +37,5 @@ function preserveDecimal(n) {
 }
 
 function preserveDecimal(n) {
-  return parseInt(n);
+  return parseFloat(n);
 }
